+++ src/shared/ui/index.ts
@@ -1,0 +1,8 @@
+export * from "./Card";
+export * from "./Button";
+export * from "./Dialog";
+export * from "./Input";
+export * from "./Textarea";
+export * from "./Input";
+export * from "./Table";
+export * from "./SelectBox";
